requirements.txt
asgiref==3.7.2
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.2.8
django-extensions==3.2.3
djangorestframework==3.14.0
executing==2.0.1
fake-factory==9999.9.9
Faker==20.1.0
ipython==8.12.3
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.0
typing-extensions==4.8.0
wcwidth==0.2.12
